+++ projects/glassui/src/lib/forms/forms.module.ts
@@ -1,16 +1,20 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {InputsModule} from "./inputs/inputs.module";
+import {SelectsModule} from "./selects/selects.module";
 
 
 @NgModule({
-  declarations: [],
+  declarations: [
+  ],
   imports: [
     CommonModule,
     InputsModule,
+    SelectsModule,
   ],
   exports: [
     InputsModule,
+    SelectsModule,
   ]
 })
 export class FormsModule {