+++ projects/glassui/src/lib/forms/selects/select-option.interface.ts
@@ -1,0 +1,5 @@
+
+export interface SelectOptionInterface{
+  value: any;
+  label: string;
+}