+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Input types (password / text)
 - Input validator
 - Inputs examples in the documentation
+- Select component
 
 ## 0.0.2-beta.2 - 2023-10-23
 ### Added